Version 10.3.0

+++ Src/PPLib/ppwtm.cpp
@@ -2340,7 +2340,7 @@ int PPWhatmanWindow::Edit(const char * pWtmFileName, const char * pWtaFileName)
 	PPWhatmanWindow * p_edit_win = 0;
 	FrameWindow * p_frame_win = new FrameWindow;
 	p_frame_win->changeBounds(TRect(10, 10, 900, 900));
-	p_frame_win->Create((long)APPL->H_MainWnd, TWindowBase::coPopup);
+	p_frame_win->Create(APPL->H_MainWnd, TWindowBase::coPopup);
 	{
 		THROW_MEM(p_tool_win = new PPWhatmanWindow(PPWhatmanWindow::modeToolbox));
 		p_frame_win->AddChild(p_tool_win, TWindowBase::coChild, FrameWindow::zoneLeft);
@@ -2456,7 +2456,7 @@ int TestFrameWindow::Run()
 	int    ok = 1;
 	TestFrameWindow * p_win = new TestFrameWindow();
 	p_win->changeBounds(TRect(10, 10, 900, 900));
-	p_win->Create((long)APPL->H_MainWnd, TWindowBase::coPopup);
+	p_win->Create(APPL->H_MainWnd, TWindowBase::coPopup);
 	{
 		TWindowBase * p_child = 0;
 		{
